Crate detcore_testutils

Testing utilities.

Macros

basic_det_test

A convenient way to wrap a function to test multiple detcore execution variants, each as a separate unit test. This is in contrast with det_test_all_configs which runs multiple modes sequentially under one test target.

make_det_test_variants

Statics

BOTTOM_CFG

Standardized test config: all options off. (This is the bottom element of a lattice containing exponentially many possibly Configs.)

MIDDLE_CFG

Standardized test config: common options on. (This is drawn from the middle of the lattice of possible Configs.)

TOP_CFG

Standardized test config: all options on. (This is the top element of a lattice containing exponentially many possibly Configs.)

Functions

check_fn_with_config

Runs a function as a guest with Detcore’s test log filter and requires a successful guest exit status.

det_test_all_configs

Runs a test across MULTIPLE configurations. This combines the multiple variants into a single test execution (running them one after another). This can be convenient at times, but, when debugging, you may want to split these apart into their own test targets.

det_test_cmd

Runs a command multiple times and checks to see if the output was deterministic between runs.

det_test_cmd_with_config

Like det_test_cmd, but allows passing in a non-default configuration.

det_test_fn

Runs a function multiple times and checks to see if the output was deterministic between runs. Expect successful exit code.

det_test_fn_with_config

Like det_test_fn, but allows passing in a non-default configuration. Takes a boolean flag indicating whether to expect a strictly deterministic behavior given the current test and current config.

det_test_fn_with_config_repetitions

Like det_test_fn_with_config, but runs and compares exactly repetitions times.

expect_success

A basic oracle, which expects a success exit code.

test_fn_with_config

Runs a function as a guest with Detcore’s test log filter applied both to the calling thread and to tracer work performed on other threads.