detcore_testutils

Function det_test_fn_with_config

pub fn det_test_fn_with_config<F, O>(
    isdet: bool,
    f: F,
    config: Config,
    oracle: O,
)
where
    F: Fn(),
    O: Fn(&Output, <Detcore as Tool>::GlobalState),

Like det_test_fn, but allows passing in a non-default configuration. Takes a boolean flag indicating whether to expect a strictly deterministic behavior given the current test and current config.