detcore_testutils

Macro basic_det_test

macro_rules! basic_det_test {
    ( $fn:path ) => { ... };
    ( $fn:path, $f:expr ) => { ... };
    ( $fn:path, $f:expr, $($variants:tt),+ ) => { ... };
    (@gendef $fn:path, $f:expr ) => { ... };
}

A convenient way to wrap a function to test multiple detcore execution variants, each as a separate unit test. This is in contrast with det_test_all_configs which runs multiple modes sequentially under one test target.

Arguments are basic_det_test(function, predicate, modes...), where:

function is the procedure under test.

predicate is a function that accepts a Config and returns true if the test should run deterministically in that configuration.

This generates calls to det_test_fn_with_config.